+++ src/modules/esm.mjs
@@ -1,0 +1,35 @@
+import path, { dirname } from 'node:path';
+import { release, version } from 'node:os';
+import { createServer as createServerHttp } from 'node:http';
+import './files/c.js';
+import { fileURLToPath } from 'node:url';
+
+const random = Math.random();
+
+const unknownObject = await import(random > 0.5 ? './files/a.json' : './files/b.json',
+    { with: { type: 'json' } });
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is '${path.sep}'`);
+
+const __filename = fileURLToPath(import.meta.url);
+console.log(`Path to current file is ${__filename}`);
+const __dirname = dirname(__filename);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknownObject, myServer };
+
